Update apoc_analysis_beast_1.md
He said Onyx, i after googling corrected to OnyxCORE. I feel based on the truth is in the message not the messenger statement from Son of Machine that message here was clearly money is a root of, and that led to the rest and the cells realization, this is like the vol 3 Alice finds Thor’s Hammer attacks / optimizations. You need ti look over time to see how this rolling alice mind control attack is based on subconscious encoding by abusing the fault tolerance inherent in the architecture.